Restructured test helpers.
